requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autopep8==2.3.2
billiard==4.2.1
build==1.2.2.post1
celery==5.5.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
django-extensions==3.2.3
django-filter==24.3
django-redis==5.4.0
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gprof2dot==2024.6.6
hiredis==3.1.0
inflection==0.5.1
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.3
Markdown==3.7
packaging==24.2
pillow==11.1.0
pip-tools==7.4.1
prompt_toolkit==3.0.51
pycodestyle==2.12.1
PyJWT==2.9.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.24.0
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
